clean up theme
