--- packages/2019-education/src/common/DataCollaborativeBrandColors.js
@@ -1,50 +1,0 @@
-export default {
-  // Plum
-  primary: {
-    rgb: "rgb(34,15,37)",
-    rgba: "rgba(34,15,37,1)",
-    hex: "#201024"
-  },
-  // Salmon
-  secondary: {
-    rgb: "rgb(220,69,86)",
-    rgba: "rgba(220,69,86,1)",
-    hex: "#DC4556"
-  },
-  // Plum
-  tertiary: {
-    rgb: "rgb(34,15,37)",
-    rgba: "rgba(34,15,37,1)",
-    hex: "#201024"
-  },
-  medium: {
-    rgb: "rgb(170,164,171)",
-    rgba: "rgba(170,164,171,1)",
-    hex: "#AAA4AB"
-  },
-  subdued: {
-    rgb: "rgb(243,242,243)",
-    rgba: "rgba(243,242,243,1)",
-    hex: "#F3F2F3"
-  },
-  action: {
-    rgb: "rgb(30,98,189)",
-    rgba: "rgba(30,98,189,1)",
-    hex: "#1E62BD"
-  },
-  background: {
-    rgb: "rgb(255,254,254)",
-    rgba: "rgba(255,254,254,1)",
-    hex: "#FFFEFE"
-  },
-  plumLight: {
-    rgb: "rgb(114,99,113)",
-    rgba: "rgba(114,99,113,1)",
-    hex: "#726371"
-  },
-  error: {
-    hex: "#DE3F8B",
-    rgb: "rgb(222,63,139)",
-    rgba: "rgb(222,63,139,1)"
-  }
-};
